JTAG uart now working, minor hw system changes * Added uart0 peripheral (RS232 Avalon UART) * Expanded juart0 buffer to 256 bytes * Regen qsys system * Fix bug with write space function * JTAG UART driver now functional and acceptable. Verified the pc connection status detection via testing with an LED. We're now getting the full greeting message out to the UART. * Will work later on getting "normal" UART working with a USB-uart adapter. This interface should be much faster, with lower latency

+++ fw/lib/inc/jtag_uart.h
@@ -35,7 +35,7 @@ typedef struct {
 void juart_init(jUartStatus* js, jUartPeriph* reg);
 void juart_serv(jUartStatus* js);
 bool juart_is_pc_conn(jUartStatus* js);
-uint8_t juart_hw_wspace(jUartStatus* js);
+uint16_t juart_hw_wspace(jUartStatus* js);
 
 void juart_write(jUartStatus* js, const char* str);
 bool juart_put(jUartStatus* js, char c);
